Begin new program to save required route data to a file for later use
This new program is designed to run ORS Directions queries using an expanded-privelege key, and save the resulting encoded polylines to a text file to be used later by the actual model. This is necessary because, in most cases, the number of routes needed by the model is more than the user's Directions query quota, so a method is needed to acquire the routes (via an expanded-qouta API key) and save them for later use.