Remove (comment out) heatmap algorithm output to prevent RAM-induced freezing
This commit was made in an effort to obtain model results before a deadline. I simply commented out the pieces of code relating to the generation of the heatmap output layer, leaving the polyline routes layer as the only output. This should alleviate the freezing/crashing issue caused by filling up my RAM with multiple copies of the heatmap layer, and I can manually process the polyline layer into a heatmap. This allows me to get the same results with only a little more work, and hopefully without freezing/crashing.